// dashboard/src/utils/hasura-api/generated/schemas/untrackFunctionBulkOperation.ts
/**
 * Generated by orval v8.4.0 🍺
 * Do not edit manually.
 * Hasura API
 * API for managing remote schemas and events in Hasura
 * OpenAPI spec version: 1.0.0
 */
import type { UntrackFunctionBulkOperationType } from './untrackFunctionBulkOperationType';
import type { UntrackFunctionStep } from './untrackFunctionStep';

export interface UntrackFunctionBulkOperation {
  type: UntrackFunctionBulkOperationType;
  source?: string;
  /** Resource version of the operation */
  resource_version: number;
  args: UntrackFunctionStep[];
}